Added FromStr implementation for Oid
